+++ Editor/Module Management/ModuleList.cs
@@ -14,7 +14,7 @@ namespace ReadyPlayerMe.Core.Editor
             name = "com.readyplayerme.core",
             gitUrl = "https://github.com/readyplayerme/rpm-unity-sdk-core.git",
             branch = "",
-            version = "0.1.1"
+            version = "0.2.0"
         };
 
         /// <summary>
@@ -34,14 +34,14 @@ namespace ReadyPlayerMe.Core.Editor
                 name = "com.readyplayerme.avatarloader",
                 gitUrl = "https://github.com/readyplayerme/rpm-unity-sdk-avatar-loader.git",
                 branch = "",
-                version = "0.1.1"
+                version = "0.2.0"
             },
             new ModuleInfo
             {
                 name = "com.readyplayerme.webview",
                 gitUrl = "https://github.com/readyplayerme/rpm-unity-sdk-webview.git",
                 branch = "",
-                version = "0.1.0"
+                version = "0.2.0"
             }
         };