+++ Editor/Module Management/ModuleUpdater.cs
@@ -1,56 +1,86 @@
 ﻿using System;
 using System.Linq;
-using UnityEditor;
-using UnityEngine;
-using Newtonsoft.Json;
 using System.Threading;
-using UnityEngine.Networking;
 using System.Threading.Tasks;
+using Newtonsoft.Json;
+using UnityEditor;
 using UnityEditor.PackageManager;
-using UnityEditor.PackageManager.Requests;
+using UnityEngine;
+using UnityEngine.Networking;
 using PackageInfo = UnityEditor.PackageManager.PackageInfo;
 
 namespace ReadyPlayerMe.Core.Editor
 {
     /// <summary>
-    ///     Class <c>ModuleUpdater</c> is responsible for checking and updating the Ready Player Me SDK modules.
+    /// It is responsible for checking and updating the Ready Player Me SDK modules.
     /// </summary>
-    public static class ModuleUpdater
+    [InitializeOnLoad]
+    public class ModuleUpdater
     {
+        private class Release
+        {
+            [JsonProperty("tag_name")]
+            public string Tag;
+        }
+
         private const string PACKAGE_JSON = "package.json";
         private const string PACKAGE_DOMAIN = "com.readyplayerme";
 
         private const string GITHUB_WEBSITE = "https://github.com";
         private const string GITHUB_API_URL = "https://api.github.com/repos";
 
-        private const string WARNING_PACKAGES_NOT_FOUND = "No Ready Player Me packages found.";
-        private const string WARNING_UPDATE_SKIPPED = "No Ready Player Me packages found.";
         private const int MILLISECONDS_TIMEOUT = 20;
         private const string ASSET_FILTER = "package";
 
+        private const string DONT_ASK = "Dont Ask";
+
+        static ModuleUpdater()
+        {
+            EntryPoint.Startup += () => Check(true);
+        }
+
         /// <summary>
-        ///     Check for Ready Player Me package updates.
+        /// Check for Ready Player Me package updates.
         /// </summary>
         [MenuItem("Ready Player Me/Check For Updates")]
         public static void CheckForUpdates()
         {
+            Check();
+        }
+
+        private static void Check(bool isStartup = false)
+        {
             // Get PackageInfo array from RPM Module package.json files
-            PackageInfo[] packages = AssetDatabase.FindAssets(ASSET_FILTER)
+            var packages = AssetDatabase.FindAssets(ASSET_FILTER)
                 .Select(AssetDatabase.GUIDToAssetPath)
                 .Where(x => x.Contains(PACKAGE_JSON) && x.Contains(PACKAGE_DOMAIN))
                 .Select(PackageInfo.FindForAssetPath)
                 .ToArray();
 
+            if (packages.Length == 0)
+            {
+                Debug.Log("No rpm package found");
+            }
+
             // Turn package_name@repo_url#branch_name into https://api.github.com/repos/readyplayerme/repo_name/releases 
-            foreach (PackageInfo package in packages)
+            foreach (var package in packages)
             {
                 var repoUrl = package.packageId.Split('@')[1];
-                var releasesUrl = repoUrl.Replace(GITHUB_WEBSITE, GITHUB_API_URL)
-                    .Split(new[] { ".git#" }, StringSplitOptions.None)[0] + "/releases";
+                var releasesUrl = repoUrl
+                    .Split(new[] { ".git" }, StringSplitOptions.None)[0] 
+                    .Replace(GITHUB_WEBSITE, GITHUB_API_URL) + "/releases";
+                
+                
                 var packageUrl = repoUrl.Split('#')[0];
 
                 // Experimental or prerelease packages might look like 0.1.0-exp.1, remove after dash to parse with Version
                 var version = package.version.Split('-')[0];
+
+                if (isStartup && EditorPrefs.GetBool(DONT_ASK + "-" + package.name))
+                {
+                    continue;
+                }
+
                 FetchReleases(package.name, packageUrl, releasesUrl, new Version(version));
             }
         }
@@ -65,18 +95,17 @@ namespace ReadyPlayerMe.Core.Editor
         private static async void FetchReleases(string packageName, string packageUrl, string releasesUrl,
             Version currentVersion)
         {
-            UnityWebRequest request = UnityWebRequest.Get(releasesUrl);
-            UnityWebRequestAsyncOperation op = request.SendWebRequest();
+            var request = UnityWebRequest.Get(releasesUrl);
+            var op = request.SendWebRequest();
             while (!op.isDone) await Task.Yield();
 
             if (request.result == UnityWebRequest.Result.Success)
             {
                 var response = request.downloadHandler.text;
-                Release[] releases = JsonConvert.DeserializeObject<Release[]>(response);
+                var releases = JsonConvert.DeserializeObject<Release[]>(response);
+                var versions = releases!.Select(r => new Version(r.Tag.Substring(1).Split('-')[0])).ToArray();
 
-                Version[] versions = releases.Select(r => new Version(r.Tag.Substring(1).Split('-')[0])).ToArray();
-
-                Version latestVersion = versions.Max();
+                var latestVersion = versions.Max();
 
                 if (latestVersion > currentVersion)
                 {
@@ -99,20 +128,27 @@ namespace ReadyPlayerMe.Core.Editor
         private static void DisplayUpdateDialog(string packageName, Version currentVersion, Version latestVersion,
             string packageUrl)
         {
-            var shouldUpdate = EditorUtility.DisplayDialog("Update Packages",
+            var shouldUpdate = EditorUtility.DisplayDialogComplex("Update Packages",
                 $"New update available for {packageName}\nCurrent version: {currentVersion}\nLatest version: {latestVersion}",
                 "Update",
-                "Skip");
+                "Cancel",
+                "Don't ask");
 
-            if (shouldUpdate)
+            switch (shouldUpdate)
             {
-                packageUrl += "#v" + latestVersion;
-                UpdateModule(packageName, packageUrl, currentVersion, latestVersion);
-            }
-            else
-            {
-                // TODO: Bring analytics here
-                Debug.LogWarning(WARNING_UPDATE_SKIPPED);
+                // Update
+                case 0:
+                    packageUrl += "#v" + latestVersion;
+                    UpdateModule(packageName, packageUrl, currentVersion, latestVersion);
+                    break;
+                // Cancel
+                case 1:
+                    // Do nothing
+                    break;
+                // Don't ask
+                case 2:
+                    EditorPrefs.SetBool(DONT_ASK + "-" + packageName, true);
+                    break;
             }
         }
 
@@ -125,19 +161,13 @@ namespace ReadyPlayerMe.Core.Editor
         /// <param name="latest">The new version of the package.</param>
         private static void UpdateModule(string name, string url, Version current, Version latest)
         {
-            RemoveRequest removeRequest = Client.Remove(name);
+            var removeRequest = Client.Remove(name);
             while (!removeRequest.IsCompleted) Thread.Sleep(MILLISECONDS_TIMEOUT);
 
-            AddRequest addRequest = Client.Add(url);
+            var addRequest = Client.Add(url);
             while (!addRequest.IsCompleted) Thread.Sleep(MILLISECONDS_TIMEOUT);
 
             Debug.Log($"Updated {name} from v{current} to v{latest}");
         }
-    }
-
-    internal class Release
-    {
-        [JsonProperty("tag_name")] 
-        public string Tag;
     }
 }