+++ CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.2.0] - 2023.02.08
+
+### Added
+- optional sdk logging
+
+### Updated
+- PartnerSubdomainSettings refactored to a CoreSettings scriptable object
+
+### Fixed
+- Various bug fixes and improvements
+
 ## [0.1.0] - 2023.01.22
 
 ### Updated